[Chore] 네이밍 변경 (#47)
- 네이밍 변경
